third_party/WebKit/LayoutTests/svg/custom/immutable-properties-expected.txt

CONSOLE WARNING: line 6: SVGSVGElement.viewport is deprecated and will be removed in M55, around November 2016. See https://www.chromestatus.com/features/5686865248124928 for more details.
Tests whether immutable properties can not be modified.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS viewport.x is 0
PASS viewport.x is 100
PASS svgDoc.documentElement.viewport.x is 0
PASS successfullyParsed is true

TEST COMPLETE
